# Optimize SimpleSubscriber for Netty