--- a/server/netty-server/src/main/scala/sttp/tapir/server/netty/internal/reactivestreams/SimpleSubscriber.scala
+++ b/server/netty-server/src/main/scala/sttp/tapir/server/netty/internal/reactivestreams/SimpleSubscriber.scala
@@ -8,13 +8,18 @@ import java.util.concurrent.ConcurrentLinkedQueue
 import scala.collection.JavaConverters._
 import scala.concurrent.{Future, Promise}
 import java.util.concurrent.LinkedBlockingQueue
+import sttp.capabilities.StreamMaxLengthExceededException
+import io.netty.buffer.Unpooled
+import io.netty.buffer.ByteBuf
+import io.netty.buffer.CompositeByteBuf
+import scala.collection.mutable
 
-private[netty] class SimpleSubscriber() extends PromisingSubscriber[Array[Byte], HttpContent] {
+private[netty] class SimpleSubscriber(contentLength: Option[Int]) extends PromisingSubscriber[Array[Byte], HttpContent] {
   private var subscription: Subscription = _
-  private val chunks = new ConcurrentLinkedQueue[Array[Byte]]()
-  private var size = 0
   private val resultPromise = Promise[Array[Byte]]()
+  private var totalLength = 0
   private val resultBlockingQueue = new LinkedBlockingQueue[Either[Throwable, Array[Byte]]]()
+  private val buffers = new ConcurrentLinkedQueue[ByteBuf]()
 
   override def future: Future[Array[Byte]] = resultPromise.future
   def resultBlocking(): Either[Throwable, Array[Byte]] = resultBlockingQueue.take()
@@ -25,44 +30,83 @@ private[netty] class SimpleSubscriber() extends PromisingSubscriber[Array[Byte],
   }
 
   override def onNext(content: HttpContent): Unit = {
-    val array = ByteBufUtil.getBytes(content.content())
-    content.release()
-    size += array.length
-    chunks.add(array)
+    val byteBuf = content.content()
+    if (contentLength.exists(_ == byteBuf.readableBytes())) {
+      val finalArray = ByteBufUtil.getBytes(byteBuf)
+      byteBuf.release()
+      resultBlockingQueue.add(Right(finalArray))
+      resultPromise.success(finalArray)
+    } else {
+      buffers.add(byteBuf)
+      totalLength += byteBuf.readableBytes()
+    }
     subscription.request(1)
   }
 
   override def onError(t: Throwable): Unit = {
-    chunks.clear()
+    buffers.forEach { buf =>
+      val _ = buf.release()
+    }
+    buffers.clear()
     resultBlockingQueue.add(Left(t))
     resultPromise.failure(t)
   }
 
   override def onComplete(): Unit = {
-    val result = new Array[Byte](size)
-    val _ = chunks.asScala.foldLeft(0)((currentPosition, array) => {
-      System.arraycopy(array, 0, result, currentPosition, array.length)
-      currentPosition + array.length
-    })
-    chunks.clear()
-    resultBlockingQueue.add(Right(result))
-    resultPromise.success(result)
+    if (!buffers.isEmpty()) {
+      val mergedArray = new Array[Byte](totalLength)
+      var currentIndex = 0
+      buffers.forEach { buf =>
+        val length = buf.readableBytes()
+        buf.getBytes(buf.readerIndex(), mergedArray, currentIndex, length)
+        currentIndex += length
+        val _ = buf.release()
+      }
+      resultBlockingQueue.add(Right(mergedArray))
+      resultPromise.success(mergedArray)
+      buffers.clear()
+    } else {
+      () // result already sent in onNext
+    }
   }
+
 }
 
 object SimpleSubscriber {
-  def processAll(publisher: Publisher[HttpContent], maxBytes: Option[Long]): Future[Array[Byte]] = {
-    val subscriber = new SimpleSubscriber()
-    publisher.subscribe(maxBytes.map(max => new LimitedLengthSubscriber(max, subscriber)).getOrElse(subscriber))
-    subscriber.future
-  }
 
-  def processAllBlocking(publisher: Publisher[HttpContent], maxBytes: Option[Long]): Array[Byte] = {
-    val subscriber = new SimpleSubscriber()
-    publisher.subscribe(maxBytes.map(max => new LimitedLengthSubscriber(max, subscriber)).getOrElse(subscriber))
-    subscriber.resultBlocking() match {
-      case Right(result) => result
-      case Left(e)       => throw e
+  def processAll(publisher: Publisher[HttpContent], contentLength: Option[Int], maxBytes: Option[Long]): Future[Array[Byte]] =
+    maxBytes match {
+      case Some(max) if (contentLength.exists(_ > max)) => Future.failed(StreamMaxLengthExceededException(max))
+      case Some(max) => {
+        val subscriber = new SimpleSubscriber(contentLength)
+        publisher.subscribe(new LimitedLengthSubscriber(max, subscriber))
+        subscriber.future
+      }
+      case None => {
+        val subscriber = new SimpleSubscriber(contentLength)
+        publisher.subscribe(subscriber)
+        subscriber.future
+      }
     }
-  }
+
+  def processAllBlocking(publisher: Publisher[HttpContent], contentLength: Option[Int], maxBytes: Option[Long]): Array[Byte] =
+    maxBytes match {
+      case Some(max) if (contentLength.exists(_ > max)) => throw new StreamMaxLengthExceededException(max)
+      case Some(max) => {
+        val subscriber = new SimpleSubscriber(contentLength)
+        publisher.subscribe(new LimitedLengthSubscriber(max, subscriber))
+        subscriber.resultBlocking() match {
+          case Right(result) => result
+          case Left(e)       => throw e
+        }
+      }
+      case None => {
+        val subscriber = new SimpleSubscriber(contentLength)
+        publisher.subscribe(subscriber)
+        subscriber.resultBlocking() match {
+          case Right(result) => result
+          case Left(e)       => throw e
+        }
+      }
+    }
 }