--- a/server/netty-server/src/main/scala/sttp/tapir/server/netty/internal/reactivestreams/SimpleSubscriber.scala
+++ b/server/netty-server/src/main/scala/sttp/tapir/server/netty/internal/reactivestreams/SimpleSubscriber.scala
@@ -1,18 +1,12 @@
 package sttp.tapir.server.netty.internal.reactivestreams
 
-import io.netty.buffer.ByteBufUtil
+import io.netty.buffer.{ByteBuf, ByteBufUtil}
 import io.netty.handler.codec.http.HttpContent
 import org.reactivestreams.{Publisher, Subscription}
-
-import java.util.concurrent.ConcurrentLinkedQueue
-import scala.collection.JavaConverters._
-import scala.concurrent.{Future, Promise}
-import java.util.concurrent.LinkedBlockingQueue
 import sttp.capabilities.StreamMaxLengthExceededException
-import io.netty.buffer.Unpooled
-import io.netty.buffer.ByteBuf
-import io.netty.buffer.CompositeByteBuf
-import scala.collection.mutable
+
+import java.util.concurrent.{ConcurrentLinkedQueue, LinkedBlockingQueue}
+import scala.concurrent.{Future, Promise}
 
 private[netty] class SimpleSubscriber(contentLength: Option[Int]) extends PromisingSubscriber[Array[Byte], HttpContent] {
   private var subscription: Subscription = _