Fixes #3548 

This PR updates the `SimpleSubscriber` with following improvements:
1. If there is only one chunk of data (by default for body length < 8192), read it from the buffer and return immediately, requiring only one array allocation
2. If there are multiple chunks, collect all Netty `ByteBufs` without rewriting them into arrays, then copy them into the final array in `onComplete`. Disclaimer: initially I wanted to cover this case with Netty's `CompositeByteBuf`, but it turned out to be bad idea. It does some reallocations to resize internal representation underneath, resulting in making the overall performance significantly worse.

Results:
`PostBytes` Simulation
|                          |      Before |  After   |
|---------------------|:-------------:|-----------:|
| CPU samples  |  30.86%   | 29.73% |
| throughput    |    44738    |   56955 |

Latency improvement:
![latency-postbytes](https://github.com/softwaremill/tapir/assets/1413553/c1dbe569-f2d6-4475-9dca-e7d2845e3674)

`PostLongBytes` Simulation
|                          |      Before |  After   |
|---------------------|:-------------:|-----------:|
| CPU samples  |  48.8%    | 38.69% |
| throughput    |    363       |  454       |

Latency improvement:
![latency-postlongbytes](https://github.com/softwaremill/tapir/assets/1413553/44d43bf8-86b3-4c93-96a1-51130e9c6296)

I haven't measured it, but there should be also noteworthy gains in memory allocations.